feat: add module for dashboard and move dashboard-view there
