+++ docs/dev-notes.md
@@ -13,7 +13,6 @@
 - consider making logo icon bigger
 - look into how easy it is to use stripe in a small form component in drawer.
 - consider adding drizzle-kit push --force to preview branches:
-- dashboard content doesn't exactly align with navbar content.
 - Add countdown selection in the navbar using a dropdown
 - Move dashboard page contents to it's own module
